Remove obsolete standard_rate page
