+++ .streamlit/config.toml
@@ -1,0 +1,7 @@
+[theme]
+base="dark"
+primaryColor="#4E79A7"
+backgroundColor="#121212"
+secondaryBackgroundColor="#1E1E1E"
+textColor="#F5F5F5"
+font="sans serif"
